! 2011-12-02 - Chris Jones  - Use long long instead of long  - Adapt to changed return type for nInputTracks()
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@132464 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 128cd0c291210276a512ccb67aeae21a34ce44e4

+++ Rec/Rich/RichRecTrackTools/src/RichDelegatedTrackCreatorFromRecoTracks.cpp
@@ -107,7 +107,7 @@ StatusCode DelegatedTrackCreatorFromRecoTracks::newTracks() const
   return StatusCode::SUCCESS;
 }
 
-long DelegatedTrackCreatorFromRecoTracks::nInputTracks() const
+unsigned int DelegatedTrackCreatorFromRecoTracks::nInputTracks() const
 {
   return ( recoTracks() ? recoTracks()->size() : 0 );
 }
